pyproject.toml
[project]
name = "llm-knowledge-base"
version = "0.1.0"
description = "LLM-powered personal knowledge base"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "anthropic>=0.40",
    "trafilatura>=1.12",
    "scikit-learn>=1.5",
    "pyyaml>=6.0",
    "httpx>=0.27",
]

[project.scripts]
kb = "kb.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["kb"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
